Version sent to Jason at Sunriver.
Fixed bug whereby if at Favorites page a user unfavoritizes an item, and that item was the last favorite in that category; and the user was previously viewing favorites at that page; then upon going back to that page, it would show no elements, and there wouldn't be a start.  Essentially the user could never view items in that category again!

+++ src/com/diamondsoftware/android/sunriver_av_3_0/DbAdapter.java
@@ -346,7 +346,7 @@ public class DbAdapter {
 						if(newVersion==24) {
 							db.execSQL("DROP TABLE IF EXISTS "+ItemHospitality.DATABASE_TABLE_HOSPITALITY);
 						} else {
-							if(newVersion==32) {
+							if(newVersion>=32 && oldVersion<32) {
 								db.execSQL("DROP TABLE IF EXISTS "+ItemService.DATABASE_TABLE_SERVICE);
 								db.execSQL("DROP TABLE IF EXISTS "+ItemWelcome.DATABASE_TABLE_WELCOME);
 							}
